connector: Move the traffic manager stuff in to a separate package
