+++ pkg/client/connector/userd_trafficmgr/install.go
@@ -1,4 +1,4 @@
-package connector
+package userd_trafficmgr
 
 import (
 	"bytes"
