+++ pkg/client/connector/userd_trafficmgr/install_actions.go
@@ -1,4 +1,4 @@
-package connector
+package userd_trafficmgr
 
 import (
 	"context"
