+++ pkg/client/connector/userd_trafficmgr/install_test.go
@@ -1,4 +1,4 @@
-package connector
+package userd_trafficmgr
 
 import (
 	"bytes"
@@ -35,7 +35,7 @@ func publishManager(t *testing.T) {
 	t.Helper()
 	ctx := dlog.NewTestContext(t, false)
 
-	cmd := dexec.CommandContext(ctx, "make", "-C", "../../..", "push-image")
+	cmd := dexec.CommandContext(ctx, "make", "-C", "../../../..", "push-image")
 
 	// Go sets a lot of variables that we don't want to pass on to the ko executable. If we do,
 	// then it builds for the platform indicated by those variables.
